Profilo test

+++ src/test/java/ta/test/impl/profile/ProfileTest.java
@@ -20,7 +20,7 @@ public class ProfileTest extends BaseTest {
     @Test(dataProvider = "fetchJSONData", dataProviderClass = JSONDataProvider.class , priority = 1)
     @Description("Test Private Profile")
 
-    public void tc_001_PrivateProfile(JSONObject testData) throws InterruptedException {
+    public void tc_001_Profile(JSONObject testData) throws InterruptedException {
 
         WebDriver driver = SeleniumDriver.getInstance().getDriver();
         driver.get(BASE_URL);
@@ -30,19 +30,29 @@ public class ProfileTest extends BaseTest {
         goLogin.clickLogin();
         loginPO.enterUsernameAndPassword(testData.get("username").toString(), testData.get("password").toString());
         loginPO.clickAccedi();
-        BrowserUtils.waitForURLContains("/mylm",30);
+        BrowserUtils.waitForURLContains("mylm", 30);
         profile.clickProfiloMattonella();
         profile.clickChipNumber();
         profile.clickCodFiscale();
         profile.clickSaveButton();
-        BrowserUtils.elementContainsText(profile.toast(),10,"Operazione completata con successo");
-        SeleniumDriver.getInstance().getDriver().manage().deleteAllCookies();
-        SessionStorage.clear();
+        try {
+            if (BrowserUtils.elementContainsText(profile.toast(), 10, "Operazione completata con successo")
+                == true) {
+                driver.navigate().to(BASE_URL+"/mylm");
+                loginPO.logout();
+                SeleniumDriver.getInstance().getDriver().manage().deleteAllCookies();
+                SessionStorage.clear();
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
     }
+
 
     @Test(dataProvider = "fetchJSONData", dataProviderClass = JSONDataProvider.class , priority=2)
     @Description("Test Company Profile")
-    public void tc_002_CompanyProfile(JSONObject testData) throws InterruptedException {
+
+    public void tc_002_Profile(JSONObject testData) throws InterruptedException {
 
         WebDriver driver = SeleniumDriver.getInstance().getDriver();
         driver.get(BASE_URL);
@@ -52,38 +62,19 @@ public class ProfileTest extends BaseTest {
         goLogin.clickLogin();
         loginPO.enterUsernameAndPassword(testData.get("username").toString(), testData.get("password").toString());
         loginPO.clickAccedi();
-        BrowserUtils.waitForURLContains("/mylm",30);
+        BrowserUtils.waitForURLContains("mylm",30);
         profile.clickProfiloMattonella();
         profile.clickChipNumber();
         profile.clickCodFiscale();
         profile.clickSaveButton();
-        BrowserUtils.elementContainsText(profile.toast(),10,"Operazione completata con successo");
-        SeleniumDriver.getInstance().getDriver().manage().deleteAllCookies();
-        SessionStorage.clear();
-    }
-
-
-
-    @Test(dataProvider = "fetchJSONData", dataProviderClass = JSONDataProvider.class , priority = 3)
-    @Description("Test Private Profile - Cambio Password")
-
-    public void tc_003_PrivateProfileChangePassword(JSONObject testData) throws InterruptedException {
-
-        WebDriver driver = SeleniumDriver.getInstance().getDriver();
-        driver.get(BASE_URL);
-        HomePagePO goLogin = new HomePagePO();
-        ProfilePO profile = new ProfilePO();
-        LoginPO loginPO = new LoginPO();
-        goLogin.clickLogin();
-        loginPO.enterUsernameAndPassword(testData.get("username").toString(), testData.get("password").toString());
-        loginPO.clickAccedi();
-        BrowserUtils.waitForURLContains("/mylm",30);
-        profile.clickProfiloMattonella();
-        profile.clickPassword();
-        profile.clickCodFiscale();
-        profile.clickSaveButton();
-        BrowserUtils.elementContainsText(profile.toast(),10,"Operazione completata con successo");
-        SeleniumDriver.getInstance().getDriver().manage().deleteAllCookies();
-        SessionStorage.clear();
+        try {
+            if (BrowserUtils.elementContainsText(profile.toast(), 10, "Operazione completata con successo")
+                == true) {
+                SeleniumDriver.getInstance().getDriver().manage().deleteAllCookies();
+                SessionStorage.clear();
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
     }
 }
